Integration: a dataset size is a Long since 6.0.0
The suite compiled against 5.x, where metadata sizes were Integers; the
published 6.0.0 answers a published file's size in 64 bits, so the test
did not compile once the registry served it.

+++ integration/src/test/java/io/vpndetection/integration/DatabaseTest.java
@@ -172,7 +172,7 @@ class DatabaseTest {
         }
         DatabaseMetadata meta = client().database().metadata(DATASET_ID);
         assertEquals(DATASET_ID, meta.getId());
-        Integer size = meta.getSize() == null ? null : meta.getSize().get(FORMAT.wireValue());
+        Long size = meta.getSize() == null ? null : meta.getSize().get(FORMAT.wireValue());
         assertNotNull(size, DATASET_ID + " publishes no size to check a transfer against");
         assertTrue(size > 0, DATASET_ID + " publishes a size of " + size);
         assertTrue(size <= CEILING, DATASET_ID + " is " + size + " bytes, past the ceiling");
